Restful conversion 1.0

+++ site/rest_ddi.py
@@ -1,9 +1,6 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
-
 """Module docstring.
 
- DNS interworking module
+ DDI API module
 
 """
 __author__ = "Zacharias El Banna"                     
@@ -11,85 +8,7 @@ __version__ = "10.0GA"
 __status__ = "Production"
 
 import sdcp.SettingsContainer as SC
-from GenLib import sys_log_msg, sys_ip2ptr, sys_ip2int, DB
-from subprocess import check_output, check_call
-from time import sleep
-from socket import gethostbyname
-
-################################ LOOPIA DNS ###################################
-#
-# Change IP for a domain in loopia DNS
-#
-loopia_domain_server_url = 'https://api.loopia.se/RPCSERV' 
-
-def set_loopia_ip(subdomain, newip):
- import xmlrpclib
- try:
-  client = xmlrpclib.ServerProxy(uri = loopia_domain_server_url, encoding = 'utf-8')
-  data = client.getZoneRecords(SC.loopia_username, SC.loopia_password, SC.loopia_domain, subdomain)[0]
-  oldip = data['rdata']
-  data['rdata'] = newip
-  status = client.updateZoneRecord(SC.loopia_username, SC.loopia_password, SC.loopia_domain, subdomain, data)[0]
- except Exception as exmlrpc:
-  sys_log_msg("System Error - Loopia set: " + str(exmlrpc))
-  return False
- return True
-
-#
-# Get Loopia settings for subdomain
-#
-def get_loopia_ip(subdomain):
- import xmlrpclib
- try:
-  client = xmlrpclib.ServerProxy(uri = loopia_domain_server_url, encoding = 'utf-8')
-  data = client.getZoneRecords(SC.loopia_username, SC.loopia_password, SC.loopia_domain, subdomain)[0]
-  return data['rdata']
- except Exception as exmlrpc:
-  sys_log_msg("System Error - Loopia get: " + str(exmlrpc))
-  return False
-
-def get_loopia_suffix():
- return "." + SC.loopia_domain
-
-################################# OpenDNS ######################################
-#
-# Return external IP from opendns
-#
-def opendns_my_ip():
- from dns import resolver
- try:
-  opendns = resolver.Resolver()
-  opendns.nameservers = [gethostbyname('resolver1.opendns.com')]
-  myiplookup = opendns.query("myip.opendns.com",'A').response.answer[0]
-  return str(myiplookup).split()[4]
- except Exception as exresolve:
-  sys_log_msg("OpenDNS Error - Resolve: " + str(exresolve))
-  return False
- 
-############################### PDNS SYSTEM FUNCTIONS ##################################
-#
-
-def pdns_get():
- recursor = check_output(["sudo","/bin/grep", "^recursor","/etc/powerdns/pdns.conf"])
- return recursor.split('=')[1].split('#')[0].strip()
-
-#
-# All-in-one, runs check, verify and (if needed) set and reload pdns service
-# - returns True if was in sync and False if modified
-# 
-def pdns_sync(dnslist):
- from XtraLib import sys_file_replace
- pdns = pdns_get()
- if not pdns in dnslist:
-  sys_log_msg("System Info - updating recursor to " + dnslist[0])
-  sys_file_replace('/etc/powerdns/pdns.conf', pdns, dnslist[0])
-  try:
-   check_call(["/usr/sbin/service","pdns","reload"])
-   sleep(1)
-  except Exception as svcerr:
-   sys_log_msg("System Error - Reloading PowerDNS: " + str(svcerr))
-  return False
- return True  
+from GenLib import sys_log_msg, sys_ip2ptr, sys_ip2int, DB, rpc_call
 
 ################################################# DDI - DNS ##################################################
 #
@@ -97,6 +16,8 @@ def pdns_sync(dnslist):
 #
 
 def ddi_dns_domains(aDict):
+ if SC.dnsdb_proxy == 'True':
+  return rpc_call(SC.dnsdb_url, "ddi_dns_domains", aDict)
  db = DB()
  db.connect_details('localhost',SC.dnsdb_username, SC.dnsdb_password, SC.dnsdb_dbname)
  res = db.do("SELECT id,name from domains")
@@ -104,6 +25,8 @@ def ddi_dns_domains(aDict):
  return retvals
 
 def ddi_dns_lookup(aDict):
+ if SC.dnsdb_proxy == 'True':
+  return rpc_call(SC.dnsdb_url, "ddi_dns_lookup", aDict)
  ptr     = sys_ip2ptr(aDict['ip'])
  fqdn    = aDict['name'] + "." + aDict['domain']
  retvals = {}
@@ -125,6 +48,8 @@ def ddi_dns_lookup(aDict):
  return retvals
 
 def ddi_dns_update(aDict):
+ if SC.dnsdb_proxy == 'True':
+  return rpc_call(SC.dnsdb_url, "ddi_dns_update", aDict)
  from time import strftime
  serial  = strftime("%Y%m%d01")
  ptr     = sys_ip2ptr(aDict['ip'])
@@ -155,6 +80,8 @@ def ddi_dns_update(aDict):
  return retvals
 
 def ddi_dns_remove(aDict):
+ if SC.dnsdb_proxy == 'True':
+  return rpc_call(SC.dnsdb_url, "ddi_dns_remove", aDict)
  db = DB()
  db.connect_details('localhost',SC.dnsdb_username, SC.dnsdb_password, SC.dnsdb_dbname)
  ares = 0
@@ -172,6 +99,8 @@ def ddi_dns_remove(aDict):
 #
 #
 def ddi_ipam_subnets(aDict):
+ if SC.ipamdb_proxy == 'True':
+  return rpc_call(SC.ipamdb_url, "ddi_ipam_subnets", aDict)
  db = DB()
  db.connect_details('localhost',SC.ipamdb_username, SC.ipamdb_password, SC.ipamdb_dbname)
  db.do("SELECT subnets.id, subnet, mask, subnets.description, name, sectionId FROM subnets INNER JOIN sections on subnets.sectionId = sections.id") 
@@ -180,6 +109,8 @@ def ddi_ipam_subnets(aDict):
  return res
 
 def ddi_ipam_lookup(aDict):
+ if SC.ipamdb_proxy == 'True':
+  return rpc_call(SC.ipamdb_url, "ddi_ipam_lookup", aDict)
  sys_log_msg("IPAM lookup - input {}".format(aDict['ip']))
  ipint   = sys_ip2int(aDict['ip'])
  retvals = { 'ipam_id':'0' }
@@ -202,6 +133,8 @@ def ddi_ipam_lookup(aDict):
 #
 #
 def ddi_ipam_update(aDict):
+ if SC.ipamdb_proxy == 'True':
+  return rpc_call(SC.ipamdb_url, "ddi_ipam_update", aDict)
  ipint = sys_ip2int(aDict['ip'])
  fqdn  = aDict['name'] + "." + aDict['domain']
  sys_log_msg("IPAM update - input:{}".format(", ".join(aDict.values())))
@@ -228,6 +161,8 @@ def ddi_ipam_update(aDict):
 #
 #
 def ddi_ipam_remove(aDict):
+ if SC.ipamdb_proxy == 'True':
+  return rpc_call(SC.ipamdb_url, "ddi_ipam_remove", aDict)
  db = DB()
  db.connect_details('localhost',SC.ipamdb_username, SC.ipamdb_password, SC.ipamdb_dbname)
  ires = db.do("DELETE FROM ipaddresses WHERE id = '{}'".format(aDict['ipam_id']))
